Correctly implement iovec versions of read/write
A regression was introduced in
f5b955fc15cd26e9b1f313bb99cce34bb2a66095
where we started handling preadv calls by the plugin with an OS
call to pread. We didn't notice the error because the tests that
use the iovec versions of read and write have not yet been
enabled.

This commit fixes it so that we are handling pread, pwrite, preadv,
and pwritev calls with the call to the corresponding OS function.